requirements.txt
Pillow
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.8.0
requests==2.27.1
urllib3==1.26.8
Werkzeug==2.0.1
